// src/styles/themes/styled.d.ts
export interface DefaultTheme {
  title: string;

  colors: {
    background: string;
    inputColor: string;
    textColor: string;
    erasedText: string,
    scndBackground: string;
    darkBackground: string;
    listBackground: string;
    primary: string;
    white: string;
    black: string;
  };
}